refactor(runner): extract environment.ts from sandbox_agent.ts
Move the environment lifecycle whole and unchanged (acquireEnvironment,
destroyInFlightSandboxes, destroyInFlightSandboxesForSession, resolveKeepaliveMount,
invalidateContinuity, plus the inFlightSandboxes registry and remount-limit const)
into environment.ts. acquireEnvironment stays a single function; the setup carve
into environment-setup.ts is a later step. Monolith re-exports the public functions
and imports acquireEnvironment/invalidateContinuity back. Pure code move; typecheck
clean; 1158 runner unit tests pass.
